Only write headers to cache folder when response has been written
See #2. To avoid any situation where headers might have been written to the
cache folder but response has not, the headers are now only written after the
response has been received and written to cache.

Note that it is, in theory, still possible to end up in a situation where
the response is written but the application gets interrupted before it has
time to write headers. This should not be a real issue as server would return
an updated response upon next conditional request.