Update assumptions in README, clarify one log message
The assumptions listed in the README were outdated as the module now acts as a
somewhat regular HTTP cache by default.

Also, one of the log messages was confusing: cache entries are not refreshed
because the cache strategy is unclear, but because there are no explicit
cache directive in the HTTP response.