Do not retry aborted requests and always handle promise rejection
When a network request is aborted through an external signal, it does not make
any sense to retry it.

Also, pending request promises were being (correctly) rejected when the request
failed, but when no one was actually interested in that request, this means that
the rejected promise went unhandled, leading to Node.js issuing an
"UnhandledPromiseRejectionWarning" message to the console.