Return 304 based on file cache info when needed
This updates the not-modified logic a bit to also return a 304 when the file
cache contains the right response and no network request gets sent (e.g. because
the `refresh` setting was set to `never`).

Handling of HTTP header names was also fixed to be case-insensitive.

This updates also introduces a couple of examples for If-None-Match and
If-Modified-Since.