[breaking] Use Node.js fetch function, drop node-fetch
This breaking change makes the library use Node.js native `fetch` function and
drops the dependency on node-fetch accordingly. Note that Node.js implementation
of `fetch` does not support cache directives, so this library remains useful for
what it was created for.

The change is breaking because it means that Node.js version 18 or greater is
now required to run the code.